Removed Kript references
Kript needs to be used in netta, and I want it to be. However, at the
moment, Kript isn't in a good state, and when trying to include it the
reference isn't found. I removed the references and re-exported the code
to make using Netta possible, however at some point I want to go back
and include the library.