Fixed bug with ConnectedClient extension
ConnectedClient was creating a problem when implemented by end-user
client. To solve this issue, I changed the parameter of ThreadAction
within MultiClientServer from ConnectedClient to Socket. This allows the
end user to take the socket, extend the ConnectedClient class, and then
pass the socket into that class, making it more streamlined for client
creation.

Side note: because of this change, the connectedClients array list is
never added to by default. It can still be used and accessed, but never
by default (unless we want to change it from type ConnectedClient to
Socket... which I dont think is best).

Re-exported docs to match changes.